\name{targetSeqView_functions}
\alias{dfmake}
\alias{dfmake2}
\alias{doPlot}
\alias{getBam}
\alias{getBamChr}
\alias{getBamPos}
\alias{getBamNames}
\alias{getBamSeqs}
\alias{getFasta}
\alias{getRandomBamRecords}
\alias{trimFunc}
\alias{withflankNs}
\alias{mainAlignView}
\title{
  Functions For targetSeqView
}
\description{
  functions for internal use
}
